PJT/backend/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
joblib==0.14.0
numpy==1.17.3
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.1
sqlparse==0.3.0
pandas==0.25.3
drf_yasg==1.17.1
django-rest-knox==4.1.0
django-cors-headers==3.2.1
mysqlclient==1.4.6 
sqlalchemy==1.3.15
pymysql==0.9.3
surprise==0.1
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
folium==0.10.1
idna==2.8
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
pandas==0.25.3
pyparsing==2.4.5
python-dateutil==2.8.1
requests==2.22.0
seaborn==0.9.0
six==1.13.0
urllib3==1.25.7
konlpy==0.5.2